+++ data-raw/regions.R
@@ -6,10 +6,11 @@ italian_region <- xml2::read_html('https://it.wikipedia.org/wiki/Regioni_d%27Ita
   dplyr::filter(Regione != 'Italia') %>%
   .[['Regione']] %>%
   stringr::str_replace_all('\\[[0-9]\\]', '') %>%
+  stringr::str_to_lower() %>%
   sort()
 
 
 regions <- list(
   italy = italian_region
 )
-use_data(regions)
+use_data(regions, overwrite = TRUE)